refactor-2: codebase refactored for simple dir structure
