--- System-Design/Design-patterns/src/com/ib/singleton/BillPughSingleton.java
@@ -1,16 +1,0 @@
-package com.ib.singleton;
-
-public class BillPughSingleton {
-    // Private constructor to prevent instantiation
-    private BillPughSingleton() {}
-    
-    // Static inner class that holds the instance
-    private static class SingletonHelper {
-        private static final BillPughSingleton INSTANCE = new BillPughSingleton();
-    }
-    
-    // Public method to get the instance
-    public static BillPughSingleton getInstance() {
-        return SingletonHelper.INSTANCE;
-    }
-}
