--- System-Design/Design-patterns/src/com/ib/singleton/DoubleCheckedSingleton.java
@@ -1,26 +1,0 @@
-package com.ib.singleton;
-
-public class DoubleCheckedSingleton {
-    // The single instance, initially null, marked as volatile
-    private static volatile DoubleCheckedSingleton instance;
-    
-    // Private constructor to prevent instantiation
-    private DoubleCheckedSingleton() {}
-    
-    // Public method to get the instance
-    public static DoubleCheckedSingleton getInstance() {
-        // First check (not synchronized)
-        if (instance == null) {
-            // Synchronize on the class object
-            synchronized (DoubleCheckedSingleton.class) {
-                // Second check (synchronized)
-                if (instance == null) {
-                    // Create the instance
-                    instance = new DoubleCheckedSingleton();
-                }
-            }
-        }
-        // Return the instance (either newly created or existing)
-        return instance;
-    }
-}
