+++ System-Design/src/design/patterns/singleton/BillPughSingleton.java
@@ -1,0 +1,16 @@
+package design.patterns.singleton;
+
+public class BillPughSingleton {
+    // Private constructor to prevent instantiation
+    private BillPughSingleton() {}
+    
+    // Static inner class that holds the instance
+    private static class SingletonHelper {
+        private static final BillPughSingleton INSTANCE = new BillPughSingleton();
+    }
+    
+    // Public method to get the instance
+    public static BillPughSingleton getInstance() {
+        return SingletonHelper.INSTANCE;
+    }
+}
